requirements.txt
anthropic>=0.40.0
notion-client>=2.2.2
exa-py>=1.1.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
